Rewrote project in C++

+++ clib/stddef.h
@@ -2,7 +2,7 @@
 #define CUSTOM_OS_CSTDDEF_H
 
 #define NULL 0
-typedef unsigned int size_t;
+typedef unsigned long size_t;
 typedef unsigned int uint;
-typedef typeof((int*) 0x00 - (int*) 0x00) ptrdiff_t; // valid since C23
+typedef typeof((int*) 0x00 - (int*) 0x00) ptrdiff_t;
 #endif //CUSTOM_OS_CSTDDEF_H
